--- a/hw10_program_optimization/new
+++ b/hw10_program_optimization/new
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/fixme_my_friend/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
+BenchmarkGetDomainStat-8   	      20	 266304418 ns/op	11867042 B/op	  646482 allocs/op
+PASS
+ok  	github.com/fixme_my_friend/hw10_program_optimization	5.725s