--- a/hw10_program_optimization/old
+++ b/hw10_program_optimization/old
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/fixme_my_friend/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
+BenchmarkGetDomainStat-8   	       4	1423382178 ns/op	307082022 B/op	 3146608 allocs/op
+PASS
+ok  	github.com/fixme_my_friend/hw10_program_optimization	13.046s